+++ packages/cxalio-studio-tools/media_killer/appenv.py
@@ -56,7 +56,7 @@ class AppEnv(IAppEnvironment):
 
         # 应用元数据
         self.app_name = "MediaKiller"
-        self.app_version = "0.9"
+        self.app_version = "0.9.1"
         self.app_description = _("媒体文件批量转码工具")
 
         # 合并自定义主题